backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.5.3
redis==5.0.1
celery==5.3.4
groq==0.4.1
openai==1.10.0
google-generativeai==0.4.0
yfinance==0.2.35
pandas==2.1.4
pillow==10.2.0
sendgrid==6.11.0
sentry-sdk==1.40.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.21.1
